[v0.90][WP-14] Split trace validation responsibilities (Closes #2050)
