+++ adl/src/cli/run_artifacts/runtime/trace_envelope.rs
@@ -1,0 +1,461 @@
+use super::super::*;
+use ::adl::trace_schema_v1::{
+    validate_trace_event_envelope_v1, ContractValidationResultV1, TraceActorTypeV1, TraceActorV1,
+    TraceContractValidationV1, TraceDecisionContextV1, TraceErrorV1, TraceEventEnvelopeV1,
+    TraceEventTypeV1, TraceEventV1, TraceScopeLevelV1, TraceScopeV1,
+};
+use serde_json::json;
+
+pub(super) fn build_trace_v1_envelope(
+    resolved: &resolve::AdlResolved,
+    tr: &trace::Trace,
+    steps: &[StepStateArtifact],
+    start_ms: u128,
+    end_ms: u128,
+    status: &str,
+    failure: Option<&anyhow::Error>,
+) -> Result<TraceEventEnvelopeV1> {
+    let mut events = Vec::new();
+    let mut next_id: u64 = 1;
+    let trace_id = resolved.run_id.clone();
+    let root_span_id = format!("run:{}", resolved.run_id);
+    let run_ref = artifact_ref(&resolved.run_id, "run.json");
+    let steps_ref = artifact_ref(&resolved.run_id, "steps.json");
+    let activation_log_ref = artifact_ref(&resolved.run_id, "logs/activation_log.json");
+
+    push_trace_v1_event(
+        &mut events,
+        &mut next_id,
+        TraceEventV1 {
+            event_id: String::new(),
+            timestamp: trace::format_iso_utc_ms(start_ms),
+            event_type: TraceEventTypeV1::RunStart,
+            trace_id: trace_id.clone(),
+            run_id: resolved.run_id.clone(),
+            span_id: root_span_id.clone(),
+            parent_span_id: None,
+            actor: TraceActorV1 {
+                r#type: TraceActorTypeV1::Agent,
+                id: resolved.workflow_id.clone(),
+            },
+            scope: TraceScopeV1 {
+                level: TraceScopeLevelV1::Run,
+                name: resolved.workflow_id.clone(),
+            },
+            inputs_ref: Some(run_ref.clone()),
+            outputs_ref: Some(activation_log_ref.clone()),
+            artifact_ref: Some(run_ref.clone()),
+            decision_context: None,
+            provider: None,
+            error: None,
+            contract_validation: None,
+        },
+    );
+
+    for event in &tr.events {
+        match event {
+            trace::TraceEvent::LifecyclePhaseEntered { ts_ms, phase, .. } => push_trace_v1_event(
+                &mut events,
+                &mut next_id,
+                TraceEventV1 {
+                    event_id: String::new(),
+                    timestamp: trace::format_iso_utc_ms(*ts_ms),
+                    event_type: TraceEventTypeV1::LifecyclePhase,
+                    trace_id: trace_id.clone(),
+                    run_id: resolved.run_id.clone(),
+                    span_id: format!("run:{}:phase:{}", resolved.run_id, phase.as_str()),
+                    parent_span_id: Some(root_span_id.clone()),
+                    actor: TraceActorV1 {
+                        r#type: TraceActorTypeV1::System,
+                        id: "runtime".to_string(),
+                    },
+                    scope: TraceScopeV1 {
+                        level: TraceScopeLevelV1::Run,
+                        name: phase.as_str().to_string(),
+                    },
+                    inputs_ref: Some(run_ref.clone()),
+                    outputs_ref: Some(activation_log_ref.clone()),
+                    artifact_ref: Some(activation_log_ref.clone()),
+                    decision_context: Some(TraceDecisionContextV1 {
+                        context: "runtime lifecycle phase".to_string(),
+                        outcome: phase.as_str().to_string(),
+                        rationale: None,
+                    }),
+                    provider: None,
+                    error: None,
+                    contract_validation: None,
+                },
+            ),
+            trace::TraceEvent::ExecutionBoundaryCrossed {
+                ts_ms,
+                boundary,
+                state,
+                ..
+            } => push_trace_v1_event(
+                &mut events,
+                &mut next_id,
+                TraceEventV1 {
+                    event_id: String::new(),
+                    timestamp: trace::format_iso_utc_ms(*ts_ms),
+                    event_type: TraceEventTypeV1::ExecutionBoundary,
+                    trace_id: trace_id.clone(),
+                    run_id: resolved.run_id.clone(),
+                    span_id: format!(
+                        "run:{}:boundary:{}:{}",
+                        resolved.run_id,
+                        boundary.as_str(),
+                        state
+                    ),
+                    parent_span_id: Some(root_span_id.clone()),
+                    actor: TraceActorV1 {
+                        r#type: TraceActorTypeV1::System,
+                        id: "runtime".to_string(),
+                    },
+                    scope: TraceScopeV1 {
+                        level: TraceScopeLevelV1::Run,
+                        name: boundary.as_str().to_string(),
+                    },
+                    inputs_ref: Some(run_ref.clone()),
+                    outputs_ref: Some(activation_log_ref.clone()),
+                    artifact_ref: Some(activation_log_ref.clone()),
+                    decision_context: Some(TraceDecisionContextV1 {
+                        context: "execution boundary".to_string(),
+                        outcome: state.clone(),
+                        rationale: None,
+                    }),
+                    provider: None,
+                    error: None,
+                    contract_validation: None,
+                },
+            ),
+            trace::TraceEvent::StepStarted {
+                ts_ms,
+                step_id,
+                agent_id,
+                ..
+            } => push_trace_v1_event(
+                &mut events,
+                &mut next_id,
+                TraceEventV1 {
+                    event_id: String::new(),
+                    timestamp: trace::format_iso_utc_ms(*ts_ms),
+                    event_type: TraceEventTypeV1::StepStart,
+                    trace_id: trace_id.clone(),
+                    run_id: resolved.run_id.clone(),
+                    span_id: format!("step:{step_id}"),
+                    parent_span_id: Some(root_span_id.clone()),
+                    actor: TraceActorV1 {
+                        r#type: TraceActorTypeV1::Agent,
+                        id: agent_id.clone(),
+                    },
+                    scope: TraceScopeV1 {
+                        level: TraceScopeLevelV1::Step,
+                        name: step_id.clone(),
+                    },
+                    inputs_ref: Some(steps_ref.clone()),
+                    outputs_ref: None,
+                    artifact_ref: Some(activation_log_ref.clone()),
+                    decision_context: None,
+                    provider: None,
+                    error: None,
+                    contract_validation: None,
+                },
+            ),
+            trace::TraceEvent::StepFinished {
+                ts_ms,
+                step_id,
+                success,
+                ..
+            } => {
+                let step_output_ref = step_artifact_ref(&resolved.run_id, steps, step_id);
+                push_trace_v1_event(
+                    &mut events,
+                    &mut next_id,
+                    TraceEventV1 {
+                        event_id: String::new(),
+                        timestamp: trace::format_iso_utc_ms(*ts_ms),
+                        event_type: TraceEventTypeV1::StepEnd,
+                        trace_id: trace_id.clone(),
+                        run_id: resolved.run_id.clone(),
+                        span_id: format!("step:{step_id}"),
+                        parent_span_id: Some(root_span_id.clone()),
+                        actor: TraceActorV1 {
+                            r#type: TraceActorTypeV1::Agent,
+                            id: resolved.workflow_id.clone(),
+                        },
+                        scope: TraceScopeV1 {
+                            level: TraceScopeLevelV1::Step,
+                            name: step_id.clone(),
+                        },
+                        inputs_ref: Some(steps_ref.clone()),
+                        outputs_ref: step_output_ref.clone().or(Some(steps_ref.clone())),
+                        artifact_ref: step_output_ref.or(Some(activation_log_ref.clone())),
+                        decision_context: None,
+                        provider: None,
+                        error: None,
+                        contract_validation: None,
+                    },
+                );
+                if !success {
+                    push_trace_v1_event(
+                        &mut events,
+                        &mut next_id,
+                        TraceEventV1 {
+                            event_id: String::new(),
+                            timestamp: trace::format_iso_utc_ms(*ts_ms),
+                            event_type: TraceEventTypeV1::Error,
+                            trace_id: trace_id.clone(),
+                            run_id: resolved.run_id.clone(),
+                            span_id: format!("step:{step_id}:error"),
+                            parent_span_id: Some(format!("step:{step_id}")),
+                            actor: TraceActorV1 {
+                                r#type: TraceActorTypeV1::System,
+                                id: "runtime".to_string(),
+                            },
+                            scope: TraceScopeV1 {
+                                level: TraceScopeLevelV1::Step,
+                                name: step_id.clone(),
+                            },
+                            inputs_ref: Some(steps_ref.clone()),
+                            outputs_ref: None,
+                            artifact_ref: Some(activation_log_ref.clone()),
+                            decision_context: None,
+                            provider: None,
+                            error: Some(TraceErrorV1 {
+                                code: "STEP_FAILURE".to_string(),
+                                message: format!("step '{step_id}' finished unsuccessfully"),
+                                details: None,
+                            }),
+                            contract_validation: None,
+                        },
+                    );
+                }
+            }
+            trace::TraceEvent::DelegationPolicyEvaluated {
+                ts_ms,
+                step_id,
+                action_kind,
+                target_id,
+                decision,
+                rule_id,
+                ..
+            } => {
+                let result = if decision.eq_ignore_ascii_case("allowed")
+                    || decision.eq_ignore_ascii_case("approved")
+                    || decision.eq_ignore_ascii_case("pass")
+                {
+                    ContractValidationResultV1::Pass
+                } else {
+                    ContractValidationResultV1::Fail
+                };
+                push_trace_v1_event(
+                    &mut events,
+                    &mut next_id,
+                    TraceEventV1 {
+                        event_id: String::new(),
+                        timestamp: trace::format_iso_utc_ms(*ts_ms),
+                        event_type: TraceEventTypeV1::ContractValidation,
+                        trace_id: trace_id.clone(),
+                        run_id: resolved.run_id.clone(),
+                        span_id: format!("step:{step_id}:policy"),
+                        parent_span_id: Some(format!("step:{step_id}")),
+                        actor: TraceActorV1 {
+                            r#type: TraceActorTypeV1::System,
+                            id: "policy-engine".to_string(),
+                        },
+                        scope: TraceScopeV1 {
+                            level: TraceScopeLevelV1::Step,
+                            name: step_id.clone(),
+                        },
+                        inputs_ref: Some(steps_ref.clone()),
+                        outputs_ref: None,
+                        artifact_ref: Some(activation_log_ref.clone()),
+                        decision_context: None,
+                        provider: None,
+                        error: None,
+                        contract_validation: Some(TraceContractValidationV1 {
+                            contract_id: "adl.delegation_policy".to_string(),
+                            result,
+                            details: Some(json!({
+                                "step_id": step_id,
+                                "action_kind": action_kind,
+                                "target_id": target_id,
+                                "decision": decision,
+                                "rule_id": rule_id,
+                            })),
+                        }),
+                    },
+                );
+            }
+            trace::TraceEvent::DelegationApproved { ts_ms, step_id, .. } => push_trace_v1_event(
+                &mut events,
+                &mut next_id,
+                TraceEventV1 {
+                    event_id: String::new(),
+                    timestamp: trace::format_iso_utc_ms(*ts_ms),
+                    event_type: TraceEventTypeV1::Approval,
+                    trace_id: trace_id.clone(),
+                    run_id: resolved.run_id.clone(),
+                    span_id: format!("step:{step_id}:approval"),
+                    parent_span_id: Some(format!("step:{step_id}")),
+                    actor: TraceActorV1 {
+                        r#type: TraceActorTypeV1::System,
+                        id: "policy-engine".to_string(),
+                    },
+                    scope: TraceScopeV1 {
+                        level: TraceScopeLevelV1::Step,
+                        name: step_id.clone(),
+                    },
+                    inputs_ref: Some(steps_ref.clone()),
+                    outputs_ref: None,
+                    artifact_ref: Some(activation_log_ref.clone()),
+                    decision_context: Some(TraceDecisionContextV1 {
+                        context: "delegation policy".to_string(),
+                        outcome: "approved".to_string(),
+                        rationale: None,
+                    }),
+                    provider: None,
+                    error: None,
+                    contract_validation: None,
+                },
+            ),
+            trace::TraceEvent::DelegationDenied {
+                ts_ms,
+                step_id,
+                action_kind,
+                target_id,
+                rule_id,
+                ..
+            } => push_trace_v1_event(
+                &mut events,
+                &mut next_id,
+                TraceEventV1 {
+                    event_id: String::new(),
+                    timestamp: trace::format_iso_utc_ms(*ts_ms),
+                    event_type: TraceEventTypeV1::Rejection,
+                    trace_id: trace_id.clone(),
+                    run_id: resolved.run_id.clone(),
+                    span_id: format!("step:{step_id}:rejection"),
+                    parent_span_id: Some(format!("step:{step_id}")),
+                    actor: TraceActorV1 {
+                        r#type: TraceActorTypeV1::System,
+                        id: "policy-engine".to_string(),
+                    },
+                    scope: TraceScopeV1 {
+                        level: TraceScopeLevelV1::Step,
+                        name: step_id.clone(),
+                    },
+                    inputs_ref: Some(steps_ref.clone()),
+                    outputs_ref: None,
+                    artifact_ref: Some(activation_log_ref.clone()),
+                    decision_context: Some(TraceDecisionContextV1 {
+                        context: format!("delegation policy {action_kind} -> {target_id}"),
+                        outcome: "denied".to_string(),
+                        rationale: rule_id.clone(),
+                    }),
+                    provider: None,
+                    error: None,
+                    contract_validation: None,
+                },
+            ),
+            trace::TraceEvent::RunFailed { ts_ms, message, .. } => push_trace_v1_event(
+                &mut events,
+                &mut next_id,
+                TraceEventV1 {
+                    event_id: String::new(),
+                    timestamp: trace::format_iso_utc_ms(*ts_ms),
+                    event_type: TraceEventTypeV1::Error,
+                    trace_id: trace_id.clone(),
+                    run_id: resolved.run_id.clone(),
+                    span_id: format!("run:{}:error", resolved.run_id),
+                    parent_span_id: Some(root_span_id.clone()),
+                    actor: TraceActorV1 {
+                        r#type: TraceActorTypeV1::System,
+                        id: "runtime".to_string(),
+                    },
+                    scope: TraceScopeV1 {
+                        level: TraceScopeLevelV1::Run,
+                        name: resolved.workflow_id.clone(),
+                    },
+                    inputs_ref: Some(run_ref.clone()),
+                    outputs_ref: None,
+                    artifact_ref: Some(activation_log_ref.clone()),
+                    decision_context: None,
+                    provider: None,
+                    error: Some(TraceErrorV1 {
+                        code: "RUN_FAILURE".to_string(),
+                        message: message.clone(),
+                        details: None,
+                    }),
+                    contract_validation: None,
+                },
+            ),
+            _ => {}
+        }
+    }
+
+    let run_end_outcome = if status == "success" {
+        "success".to_string()
+    } else if status == "paused" {
+        "paused".to_string()
+    } else {
+        "failure".to_string()
+    };
+    push_trace_v1_event(
+        &mut events,
+        &mut next_id,
+        TraceEventV1 {
+            event_id: String::new(),
+            timestamp: trace::format_iso_utc_ms(end_ms),
+            event_type: TraceEventTypeV1::RunEnd,
+            trace_id,
+            run_id: resolved.run_id.clone(),
+            span_id: root_span_id,
+            parent_span_id: None,
+            actor: TraceActorV1 {
+                r#type: TraceActorTypeV1::Agent,
+                id: resolved.workflow_id.clone(),
+            },
+            scope: TraceScopeV1 {
+                level: TraceScopeLevelV1::Run,
+                name: resolved.workflow_id.clone(),
+            },
+            inputs_ref: Some(run_ref),
+            outputs_ref: Some(steps_ref),
+            artifact_ref: Some(activation_log_ref),
+            decision_context: Some(TraceDecisionContextV1 {
+                context: "run completion".to_string(),
+                outcome: run_end_outcome,
+                rationale: failure.map(|err| err.to_string()),
+            }),
+            provider: None,
+            error: None,
+            contract_validation: None,
+        },
+    );
+
+    let envelope = TraceEventEnvelopeV1 {
+        schema_version: "trace.v1".to_string(),
+        events,
+    };
+    validate_trace_event_envelope_v1(&envelope)?;
+    Ok(envelope)
+}
+
+fn push_trace_v1_event(events: &mut Vec<TraceEventV1>, next_id: &mut u64, mut event: TraceEventV1) {
+    event.event_id = format!("trace-v1-{:04}", *next_id);
+    *next_id = next_id.saturating_add(1);
+    events.push(event);
+}
+
+fn artifact_ref(run_id: &str, relative_path: &str) -> String {
+    format!("artifacts/{run_id}/{relative_path}")
+}
+
+fn step_artifact_ref(run_id: &str, steps: &[StepStateArtifact], step_id: &str) -> Option<String> {
+    let rel = steps
+        .iter()
+        .find(|step| step.step_id == step_id)
+        .and_then(|step| step.output_artifact_path.as_deref())?;
+    Some(artifact_ref(run_id, rel))
+}
